% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_calc.R
\name{prs_calc}
\alias{prs_calc}
\title{Calculate Polygenic Risk Score}
\usage{
prs_calc(base_file = NULL, base_header = TRUE, target_bed = NULL)
}
\arguments{
\item{base_file}{path and file name to summary statistics}

\item{base_header}{a logical value indicating whether the first row of base_file contains column names, default is TRUE.}

\item{target_bed}{path and file name to plink's .bed individual-level genotype-phenotype
file, .bim and .fam files must also be located in the same directory with the same file name,
prefix, e.g. EUR.bed, EUR.bim, EUR.fam.}
}
\value{
a numeric
}
\description{
Calculate Polygenic Risk Score
}
\examples{
prs_calc(base_file = "./data/extdat/Height.QC.Transformed", target_bed = "./inst/extdat/EUR.QC.bed")
}
